requirements.txt
pytest
pytest-html
flake8
psycopg2-binary
pymongo
